added a new Entity for historicalData and refactored all my Entities to use Long as ID

+++ springboot-trading-data/src/main/java/de/segoy/springboottradingdata/modelconverter/DatabaseSyncIBKRContractAndContractData.java
@@ -7,7 +7,7 @@ import de.segoy.springboottradingdata.repository.ComboLegDataRepository;
 import de.segoy.springboottradingdata.repository.ContractDataRepository;
 import org.springframework.stereotype.Component;
 
-import java.util.OptionalInt;
+import java.util.OptionalLong;
 
 @Component
 public class DatabaseSyncIBKRContractAndContractData {
@@ -22,7 +22,7 @@ public class DatabaseSyncIBKRContractAndContractData {
         this.comboLegDataRepository = comboLegDataRepository;
     }
 
-    public ContractData findInDBOrConvertAndSaveOrUpdateIfIdIsProvided(OptionalInt id, Contract contract){
+    public ContractData findInDBOrConvertAndSaveOrUpdateIfIdIsProvided(OptionalLong id, Contract contract){
         return contractDataRepository.findFirstByContractId(contract.conid())
                 .orElseGet(()->{
             ContractData newContractData = ibkrContractToContractData.convertIBKRContract(contract);
