+++ springboot-trading-data/src/main/java/de/segoy/springboottradingdata/repository/BaseRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.NoRepositoryBean;
 
 @NoRepositoryBean
-public interface BaseRepository<T extends BaseEntity> extends CrudRepository<T, Integer> {
+public interface BaseRepository<T extends BaseEntity> extends CrudRepository<T, Long> {
 }
